requirements.txt
numpy
matplotlib
opencv-python
tensorflow
cython
Pillow